# backend/srcs/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST du projet Pan Bagnat.",
        "title": "Pan Bagnat API",
        "contact": {},
        "version": "1.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/modules/{moduleID}": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Returns all information about a module given its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Modules"
                ],
                "summary": "Get Module",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Module ID",
                        "name": "moduleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Module"
                        }
                    },
                    "400": {
                        "description": "ID not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates the specified fields of a module (e.g. name, Git URL/branch, status).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Modules"
                ],
                "summary": "Patch Module",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Module ID",
                        "name": "moduleID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modules.ModulePatchInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The updated module",
                        "schema": {
                            "$ref": "#/definitions/api.Module"
                        }
                    },
                    "400": {
                        "description": "Invalid module ID or JSON body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Module not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/admin/modules/{moduleID}/pages/{pageID}": {
            "delete": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Delete a module page for your campus (All page datas will be lost!)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Modules",
                    "Pages"
                ],
                "summary": "Delete Module page",
                "parameters": [
                    {
                        "description": "Module input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modules.ModulePatchInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "description": "Updates the metadata of a module’s page (name, URL, visibility).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Modules",
                    "Pages"
                ],
                "summary": "Patch Module Page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Module ID",
                        "name": "moduleID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Page ID",
                        "name": "pageID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modules.ModulePageUpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The updated module page",
                        "schema": {
                            "$ref": "#/definitions/api.ModulePage"
                        }
                    },
                    "400": {
                        "description": "Bad request (missing or invalid parameters)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Page not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Module": {
            "type": "object",
            "properties": {
                "git_branch": {
                    "description": "GitBranch is the branch currently tracked for deployments",
                    "type": "string",
                    "example": "main"
                },
                "git_url": {
                    "description": "GitURL is the repository URL where the module’s source lives",
                    "type": "string",
                    "example": "https://github.com/some-user/some-repo"
                },
                "icon_url": {
                    "description": "IconURL is the link to the module’s icon or logo",
                    "type": "string",
                    "example": "https://someURL/image.png"
                },
                "id": {
                    "description": "ID is the unique identifier of the module",
                    "type": "string",
                    "example": "module_01HZ0MMK4S6VQW4WPHB6NZ7R7X"
                },
                "last_update": {
                    "description": "LastUpdate is the UTC timestamp of the module’s last update check",
                    "type": "string",
                    "example": "2025-02-18T15:00:00Z"
                },
                "late_commits": {
                    "description": "LateCommits is the number of commits behind the latest version",
                    "type": "integer",
                    "example": 2
                },
                "latest_version": {
                    "description": "LatestVersion is the most recent version available upstream",
                    "type": "string",
                    "example": "1.7"
                },
                "name": {
                    "description": "Name is the human-readable title of the module",
                    "type": "string",
                    "example": "Captain Hook"
                },
                "roles": {
                    "description": "Roles lists the roles that have access to this module",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Role"
                    }
                },
                "slug": {
                    "description": "Slug is the URL-friendly identifier for the module",
                    "type": "string",
                    "example": "captain-hook-main"
                },
                "ssh_public_key": {
                    "description": "SSHPublicKey is the RSA public key used for module deployments",
                    "type": "string",
                    "example": "ssh-rsa AAAA..."
                },
                "status": {
                    "description": "Status indicates whether the module is enabled or disabled",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.ModuleStatus"
                        }
                    ],
                    "example": "enabled"
                },
                "version": {
                    "description": "Version is the currently deployed version of the module",
                    "type": "string",
                    "example": "1.2"
                }
            }
        },
...
        "api.ModulePage": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "page_01HZ0MMK4S6VQW4WPHB6NZ7R7X"
                },
                "is_public": {
                    "type": "boolean"
                },
                "module_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "api.ModuleStatus": {
            "type": "string",
            "enum": [
                "cloning",
                "waiting_for_deploy",
                "disabled",
                "enabled"
            ],
            "x-enum-varnames": [
                "Cloning",
                "WaitingDeploy",
                "Disabled",
                "Enabled"
            ]
        },
        "api.Role": {
            "type": "object",
            "properties": {
                "color": {
                    "description": "Color is the hex color code associated with the role in the UI",
                    "type": "string",
                    "example": "#FF00FF"
                },
                "id": {
                    "description": "ID is the unique identifier of the role",
                    "type": "string",
                    "example": "role_01HZ0MMK4S6VQW4WPHB6NZ7R7X"
                },
                "is_default": {
                    "description": "IsDefault indicates whether this role is assigned by default to new users",
                    "type": "boolean",
                    "example": true
                },
                "modules": {
                    "description": "Modules lists the Module objects this role grants access to",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Module"
                    }
                },
                "name": {
                    "description": "Name is the human‐readable name of the role (e.g. \"IT\", \"Admin\")",
                    "type": "string",
                    "example": "IT"
                },
                "users": {
                    "description": "Users lists the User objects that have this role",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.User"
                    }
                },
                "usersCount": {
                    "description": "UsersCount is the total number of users assigned to this role",
                    "type": "integer",
                    "example": 42
                }
            }
        },
        "api.User": {
            "type": "object",
            "properties": {
                "ft_id": {
                    "description": "FtID is the numeric 42-intranet ID of the user",
                    "type": "integer",
                    "example": 1492
                },
                "ft_is_staff": {
                    "description": "FtIsStaff indicates whether the user is a 42-intranet staff member",
                    "type": "boolean",
                    "example": true
                },
                "ft_login": {
                    "description": "FtLogin is the 42-intranet login handle (e.g. \"heinz\")",
                    "type": "string",
                    "example": "heinz"
                },
                "ft_photo": {
                    "description": "PhotoURL is the URL to the user’s 42-intranet profile picture",
                    "type": "string",
                    "example": "https://intra.42.fr/some-login/some-id"
                },
                "id": {
                    "description": "ID is the unique identifier of the user",
                    "type": "string",
                    "example": "user_01HZ0MMK4S6VQW4WPHB6NZ7R7X"
                },
                "is_staff": {
                    "description": "IsStaff indicates whether the user has staff privileges within Pan Bagnat",
                    "type": "boolean",
                    "example": true
                },
                "last_seen": {
                    "description": "LastSeen is the UTC timestamp of the user’s last activity",
                    "type": "string",
                    "example": "2025-02-18T15:00:00Z"
                },
                "roles": {
                    "description": "Roles lists the roles assigned to the user",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Role"
                    }
                }
            }
        },
...
        "modules.ModulePageUpdateInput": {
            "type": "object",
            "properties": {
                "is_public": {
                    "description": "IsPublic toggles whether the page is public.",
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "description": "Name is the new name for the page.",
                    "type": "string",
                    "example": "Home"
                },
                "url": {
                    "description": "URL is the new URL for the page.",
                    "type": "string",
                    "example": "https://example.com/home"
                }
            }
        },
...
        "modules.ModulePatchInput": {
            "description": "API Module model",
            "type": "object",
            "properties": {
                "gitBranch": {
                    "type": "string",
                    "example": "main"
                },
                "git_url": {
                    "type": "string",
                    "example": "https://github.com/some-user/some-repo"
                },
                "name": {
                    "type": "string",
                    "example": "captain-hook"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminAuth": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        },
        "SessionAuth": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    },
...
}